+++ src/textual/app.py
@@ -1319,7 +1319,9 @@ class App(Generic[ReturnType], DOMNode):
             keys, action, description, show=show, key_display=key_display
         )
 
-    def get_key_display(self, key: str) -> str:
+    def get_key_display(
+        self, key: str, upper_case_keys: bool = False, ctrl_to_caret: bool = True
+    ) -> str:
         """For a given key, return how it should be displayed in an app
         (e.g. in the Footer widget).
         By key, we refer to the string used in the "key" argument for
@@ -1329,11 +1331,15 @@ class App(Generic[ReturnType], DOMNode):
 
         Args:
             key: The binding key string.
+            upper_case_keys: Upper case printable keys.
+            ctrl_to_caret: Replace `ctrl+` with `^`.
 
         Returns:
             The display string for the input key.
         """
-        return _get_key_display(key)
+        return _get_key_display(
+            key, upper_case_keys=upper_case_keys, ctrl_to_caret=ctrl_to_caret
+        )
 
     async def _press_keys(self, keys: Iterable[str]) -> None:
         """A task to send key events."""
@@ -3497,6 +3503,19 @@ class App(Generic[ReturnType], DOMNode):
     def action_focus_previous(self) -> None:
         """An [action](/guide/actions) to focus the previous widget."""
         self.screen.focus_previous()
+
+    def action_hide_keys(self) -> None:
+        """Hide the keys panel (if present)."""
+        self.screen.query("KeyPanel").remove()
+
+    def action_show_keys(self) -> None:
+        """Show the keys panel."""
+        from .widgets import KeyPanel
+
+        try:
+            self.query_one(KeyPanel)
+        except NoMatches:
+            self.mount(KeyPanel())
 
     def _on_terminal_supports_synchronized_output(
         self, message: messages.TerminalSupportsSynchronizedOutput