+++ src/textual/keys.py
@@ -259,10 +259,12 @@ KEY_DISPLAY_ALIASES = {
     "left": "←",
     "right": "→",
     "backspace": "⌫",
-    "escape": "ESC",
+    "escape": "esc",
     "enter": "⏎",
     "minus": "-",
-    "space": "SPACE",
+    "space": "space",
+    "pagedown": "pgdn",
+    "pageup": "pgup",
 }
 
 
@@ -279,12 +281,24 @@ def _get_key_aliases(key: str) -> list[str]:
     return [key] + KEY_ALIASES.get(key, [])
 
 
-def _get_key_display(key: str) -> str:
+def _get_key_display(
+    key: str,
+    upper_case_keys: bool = False,
+    ctrl_to_caret: bool = True,
+) -> str:
     """Given a key (i.e. the `key` string argument to Binding __init__),
     return the value that should be displayed in the app when referring
     to this key (e.g. in the Footer widget)."""
     if "+" in key:
-        return "+".join([_get_key_display(key) for key in key.split("+")])
+        key_components = key.split("+")
+        caret = False
+        if ctrl_to_caret and "ctrl" in key_components:
+            key_components.remove("ctrl")
+            caret = True
+        key_display = ("^" if caret else "") + "+".join(
+            [_get_key_display(key) for key in key_components]
+        )
+        return key_display
 
     display_alias = KEY_DISPLAY_ALIASES.get(key)
     if display_alias:
@@ -300,7 +314,7 @@ def _get_key_display(key: str) -> str:
     # Check if printable. `delete` for example maps to a control sequence
     # which we don't want to write to the terminal.
     if unicode_character.isprintable():
-        return unicode_character
+        return unicode_character.upper() if upper_case_keys else unicode_character
     return tentative_unicode_name