+++ tests/snapshot_tests/test_snapshots.py
@@ -1438,3 +1438,8 @@ def test_command_palette_dismiss_escape_no_results(snap_compare):
     assert snap_compare(
         SNAPSHOT_APPS_DIR / "command_palette_dismiss.py", run_before=run_before
     )
+
+
+def test_split(snap_compare):
+    """Test split rule."""
+    assert snap_compare(SNAPSHOT_APPS_DIR / "split.py", terminal_size=(100, 30))